+++ src/expense_forecast/InvestmentBillingState.py
@@ -1,15 +1,14 @@
 from dataclasses import dataclass
 from datetime import date
-from decimal import Decimal
 
 
 @dataclass
 class InvestmentBillingState:
 
     def __init__(self, 
-                balance: Decimal,
+                balance: float,
                 billing_start_date: date, 
-                expected_apr: Decimal
+                expected_apr: float
                 ):
         """
         Represents the billing state for an investment account.
@@ -21,15 +20,15 @@ class InvestmentBillingState:
 
         Parameters
         ----------
-        balance : Decimal
+        balance : float
             Current account balance at the beginning of the billing cycle.
 
         billing_cycle_start_date : date
             Date on which the current billing or accrual cycle begins.
 
-        expected_apr : Decimal
+        expected_apr : float
             Expected annual percentage return expressed as a decimal fraction.
-            For example, use ``Decimal("0.07")`` for an expected annual return
+            For example, use ``float("0.07")`` for an expected annual return
             of 7%.
 
 
@@ -37,8 +36,8 @@ class InvestmentBillingState:
         """
         self.billing_cycle_start_date = billing_start_date
         
-        balance = Decimal(str(balance))
-        expected_apr = Decimal(str(expected_apr))
+        balance = float(str(balance))
+        expected_apr = float(str(expected_apr))
 
         assert balance >= 0
         self.balance = balance
@@ -46,11 +45,11 @@ class InvestmentBillingState:
         assert expected_apr >= 0
         self.apr = expected_apr
 
-    def accrue_return(self, days=1) -> Decimal:
+    def accrue_return(self, days=1) -> float:
         """Compound the balance daily for ``days`` and return the growth."""
         if days < 0:
             raise ValueError("Investment accrual days cannot be negative")
         starting_balance = self.balance
-        daily_growth_factor = Decimal("1") + self.apr / Decimal("365.25")
+        daily_growth_factor = float("1") + self.apr / float("365.25")
         self.balance *= daily_growth_factor ** days
         return self.balance - starting_balance
